composition_advisor/pyproject.toml
[project]
name = "composition-advisor"
version = "0.1.0"
description = "MIDI composition critique tool for jazz/fusion/city-pop/gospel writers."
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.91.0",
    "music21>=9.9.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "typer>=0.24.1",
]

[project.scripts]
analyze = "composition_advisor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["composition_advisor"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
